Split package.rs into several files
